Renderer code clean up and misc tweaks
* Imagine: Simplify handling of common renderer programs
* Imagine: Set initial GL context state after RenderTask is created instead of in RendererCommands
* Imagine: Move GLTask semaphore management into its delegates
* Imagine: Add destructor and move constructor to Gfx::DrawableHolder
* Imagine: Add Gfx::Projection class to store a projection plane and matrix pair
* Imagine: Add Gfx::ShaderType enum
* Imagine: Use a std::shared_ptr to store Window class custom data
* Imagine: Add IG_enableMemberIf/IG_enableMemberIfOrConstant macros
* Imagine: Initialize X11 Window position values in XCreateWindow()/XSetWMNormalHints()
* Imagine: Replace bitsSet() function with C++20 std::popcount()
* EmuFramework: Minimize global access of vController

+++ imagine/include/imagine/base/baseDefs.hh
@@ -92,7 +92,7 @@ using namespace IG;
 using OnFrameDelegate = DelegateFunc<bool (FrameParams params)>;
 
 // orientation
-using Orientation = uint32_t;
+using Orientation = uint8_t;
 static constexpr Orientation VIEW_ROTATE_0 = bit(0), VIEW_ROTATE_90 = bit(1), VIEW_ROTATE_180 = bit(2), VIEW_ROTATE_270 = bit(3);
 static constexpr Orientation VIEW_ROTATE_AUTO = bit(5);
 static constexpr Orientation VIEW_ROTATE_ALL = VIEW_ROTATE_0 | VIEW_ROTATE_90 | VIEW_ROTATE_180 | VIEW_ROTATE_270;
